+++ legends/ui/__init__.py
@@ -10,6 +10,8 @@ Example:
 
 """
 
+from legends.ui.charcard import *
 from legends.ui.dialogs import *
 from legends.ui.rostertab import *
+from legends.ui.session import *
 from legends.ui.stlplanner import *